Move calculateMaxComponent to MediaControls
