+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -32,7 +32,7 @@ import $L from '../internal/$L';
 import Spinner from '../Spinner';
 import Skinnable from '../Skinnable';
 
-import {calcNumberValueOfPlaybackRate, compareSources, countReactChildren, secondsToTime} from './util';
+import {calcNumberValueOfPlaybackRate, compareSources, secondsToTime} from './util';
 import Overlay from './Overlay';
 import MediaControls from './MediaControls';
 import MediaTitle from './MediaTitle';
@@ -649,27 +649,9 @@ const VideoPlayerBase = class extends React.Component {
 		on('keydown', this.handleGlobalKeyDown);
 		on('keyup', this.handleKeyUp);
 		this.startDelayedFeedbackHide();
-		this.calculateMaxComponentCount(
-			countReactChildren(this.props.leftComponents),
-			countReactChildren(this.props.rightComponents),
-			countReactChildren(this.props.children)
-		);
 	}
 
 	componentWillReceiveProps (nextProps) {
-		// Detect if the number of components has changed
-		const leftCount = countReactChildren(nextProps.leftComponents),
-			rightCount = countReactChildren(nextProps.rightComponents),
-			childrenCount = countReactChildren(nextProps.children);
-
-		if (
-			countReactChildren(this.props.leftComponents) !== leftCount ||
-			countReactChildren(this.props.rightComponents) !== rightCount ||
-			countReactChildren(this.props.children) !== childrenCount
-		) {
-			this.calculateMaxComponentCount(leftCount, rightCount, childrenCount);
-		}
-
 		const {source} = this.props;
 		const {source: nextSource} = nextProps;
 
@@ -790,17 +772,6 @@ const VideoPlayerBase = class extends React.Component {
 		} else {
 			return 0;
 		}
-	}
-
-	calculateMaxComponentCount = (leftCount, rightCount, childrenCount) => {
-		// If the "more" button is present, automatically add it to the right's count.
-		if (childrenCount) {
-			rightCount += 1;
-		}
-
-		const max = Math.max(leftCount, rightCount);
-
-		this.player.style.setProperty('--moon-video-player-max-side-components', max);
 	}
 
 	initI18n = () => {
